+++ core/TestRacer.ts
@@ -64,7 +64,8 @@ class TestRacerTestBlock {
     this._emitMessage(
       failures
         .map((failure: any) => {
-          const actual = new jtree.TreeNode(`actual\n${new jtree.TreeNode(failure[0].toString()).toString(1)}`)
+          const actualVal = failure[0] === undefined ? "undefined" : failure[0].toString()
+          const actual = new jtree.TreeNode(`actual\n${new jtree.TreeNode(actualVal).toString(1)}`)
           const expected = new jtree.TreeNode(`expected\n${new jtree.TreeNode(failure[1].toString()).toString(1)}`)
           const comparison = actual.toComparison(expected)
           return new jtree.TreeNode(` assertion ${failure[2]}\n${comparison.toSideBySide([actual, expected]).toString(2)}`)